[change] data & price for more flexibility

+++ internal/product/entity/product.go
@@ -1,8 +1,8 @@
 package entity
 
 type Product struct {
-	ID    int    `json:"id"`
-	Name  string `json:"name"`
-	Qty   int    `json:"qty"`
-	Price int    `json:"price"`
+	ID    int     `json:"id"`
+	Name  string  `json:"name"`
+	Qty   int     `json:"qty"`
+	Price float64 `json:"price"`
 }
